+++ tests/unittests/app.py
@@ -82,20 +82,3 @@ def test_app_grammar(patch):
     patch.init(Grammar)
     patch.object(Grammar, 'build')
     assert App.grammar() == Grammar().build()
-
-
-def test_app_loads(patch):
-    patch.init(Story)
-    patch.object(Story, 'process')
-    result = App.loads('string')
-    Story.__init__.assert_called_with('string')
-    assert result == Story.process()
-
-
-def test_app_load(patch, magic):
-    patch.object(Story, 'from_stream')
-    stream = magic()
-    result = App.load(stream)
-    Story.from_stream.assert_called_with(stream)
-    story = Story.from_stream().process()
-    assert result == {stream.name: story, 'services': story['services']}
